[New] organize code into seperate files
