+++ src/base/browser/secondary/listWidget/listWidget.ts
@@ -1,18 +1,17 @@
+import { IListItemProvider } from "src/base/browser/secondary/listView/listItemProvider";
 import { IListViewRenderer, PipelineRenderer } from "src/base/browser/secondary/listView/listRenderer";
 import { IListViewOpts, IViewItemChangeEvent, ListError, ListView } from "src/base/browser/secondary/listView/listView";
-import { DisposableManager, IDisposable } from "src/base/common/dispose";
-import { DomUtility } from "src/base/browser/basic/dom";
-import { Event, Register, SignalEmitter } from "src/base/common/event";
-import { IScrollEvent } from "src/base/common/scrollable";
-import { IListItemProvider } from "src/base/browser/secondary/listView/listItemProvider";
 import { IListDragAndDropProvider, ListWidgetDragAndDropController } from "src/base/browser/secondary/listWidget/listWidgetDragAndDrop";
-import { memoize } from "src/base/common/memoization";
-import { Arrays } from "src/base/common/util/array";
-import { IS_MAC } from "src/base/common/platform";
-import { createStandardKeyboardEvent, IStandardKeyboardEvent, KeyCode } from "src/base/common/keyboard";
-import { IRange } from "src/base/common/range";
-import { isNumber, NulltoUndefined } from "src/base/common/util/type";
+import { ListWidgetKeyboardController } from "src/base/browser/secondary/listWidget/listWidgetKeyboardController";
+import { ListWidgetMouseController } from "src/base/browser/secondary/listWidget/listWidgetMouseController";
 import { ListTrait, ListTraitRenderer } from "src/base/browser/secondary/listWidget/listWidgetTrait";
+import { DisposableManager, IDisposable } from "src/base/common/dispose";
+import { Event, Register, SignalEmitter } from "src/base/common/event";
+import { createStandardKeyboardEvent, IStandardKeyboardEvent, KeyCode } from "src/base/common/keyboard";
+import { memoize } from "src/base/common/memoization";
+import { IRange } from "src/base/common/range";
+import { IScrollEvent } from "src/base/common/scrollable";
+import { isNumber, NulltoUndefined } from "src/base/common/util/type";
 
 /**
  * The index changed in {@link ListTrait}.
@@ -21,333 +20,6 @@ export interface ITraitChangeEvent {
 
     /** The new indices with the corresponding trait. */
     indice: number[];
-}
-
-/**
- * @internal
- * @class An internal class that handles the mouse support of {@link IListWidget}.
- * It handles:
- *  - when to focus DOM
- *  - when to focus item
- *  - when to select item(s)
- * 
- * @readonly EXPORT FOR OTHER MODULES ONLY. DO NOT USE DIRECTLY.
- */
-export class ListWidgetMouseController<T> implements IDisposable {
-
-    // [fields]
-
-    private _disposables = new DisposableManager();
-    private _view: IListWidget<T>;
-
-    private _multiSelectionSupport: boolean = true;
-
-    // [constructor]
-
-    constructor(
-        view: IListWidget<T>,
-        opts: IListWidgetOpts<T>
-    ) {
-        this._view = view;
-
-        this._view.DOMElement.classList.add('mouse-support');
-
-        if (opts.multiSelectionSupport !== undefined) {
-            this._multiSelectionSupport = opts.multiSelectionSupport;
-        }
-
-        this._disposables.register(view.onMousedown(e => this.__onMouseDown(e)));
-        this._disposables.register(view.onClick(e => this.__onMouseClick(e)));
-    }
-
-    // [public methods]
-
-    public dispose(): void {
-        this._disposables.dispose();
-    }
-
-    // [protect methods]
-
-    protected __ifSupported(e: IListMouseEvent<T>): boolean {
-        if (DomUtility.isInputElement(e.browserEvent.target as HTMLElement)) {
-            return false;
-        }
-        return true;
-    }
-
-    /**
-     * @description Handles item focus and selection logic.
-     */
-    protected __onMouseClick(e: IListMouseEvent<T>): void {
-
-        if (this.__ifSupported(e) === false) {
-            return;
-        }
-
-        const toFocused = e.actualIndex;
-        
-        // clicking nowhere, we reset all the traits
-        if (toFocused === undefined) {
-            this._view.setFocus(null);
-            this._view.setAnchor(null);
-            this._view.setSelections([]);
-            return;
-        }
-
-        // check if selecting in range
-        if (this.__isSelectingInRangeEvent(e)) {
-            this.__multiSelectionInRange(e);
-            return;
-        } else if (this.__isSelectingInSingleEvent(e)) {
-            this._mutliSelectionInSingle(e);
-            return;
-        }
-
-        // normal click
-        this._view.setAnchor(toFocused);
-        this._view.setFocus(toFocused);
-        if (DomUtility.isMouseRightClick(e.browserEvent) === false) {
-            this._view.setSelections([toFocused]);
-        }
-    }
-
-    /**
-     * @description Determines if the event is selecting in range. In other words,
-     * pressing SHIFT.
-     */
-    protected __isSelectingInRangeEvent(e: IListMouseEvent<T>): boolean {
-        if (this._multiSelectionSupport === false) {
-            return false;
-        }
-        return e.browserEvent.shiftKey;
-    }
-
-    /**
-     * @description Determines if the event is selecting in single. In other words,
-     * pressing CTRL in Windows or META in Macintosh.
-     */
-    protected __isSelectingInSingleEvent(e: IListMouseEvent<T>): boolean {
-        if (this._multiSelectionSupport === false) {
-            return false;
-        }
-        return IS_MAC ? e.browserEvent.metaKey : e.browserEvent.ctrlKey;
-    }
-
-    // [private helper methods]
-
-    /**
-     * @description Focuses the event target element.
-     */
-    private __onMouseDown(e: IListMouseEvent<T>): void {
-        // prevent double focus
-        if (document.activeElement !== e.browserEvent.target) {
-			this._view.setDomFocus();
-		}
-    }
-
-    /**
-     * @description Applies multi-selection when selecting in range.
-     */
-    private __multiSelectionInRange(e: IListMouseEvent<T>): void {
-        const toFocused = e.actualIndex!;
-        let anchor = this._view.getAnchor();
-
-        // if no focus yet, we focus on the current.
-        if (anchor === null) {
-            anchor = this._view.getFocus() ?? toFocused;
-            this._view.setAnchor(anchor);
-        }
-
-        /**
-         * @readonly Below is not really a good implementation (could be optimized), 
-         * but works.
-         */
-
-        // calculates the selection range
-        const toSelectRange = Arrays.range(
-            Math.min(toFocused, anchor), 
-            Math.max(toFocused, anchor) + 1
-        );
-        const currSelection = this._view.getSelections().sort((a, b) => a - b);
-        const contiguousRange = this.__getNearestContiguousRange(Arrays.unique(Arrays.insert(currSelection, anchor)), anchor);
-        if (!contiguousRange.length) {
-            return;
-        }
-        const newSelection = Arrays.union(toSelectRange, 
-                                Arrays.union(
-                                    Arrays.relativeComplement(currSelection, contiguousRange), 
-                                    Arrays.relativeComplement(contiguousRange, currSelection)
-                                )
-                            );
-        
-        // update selections and focused
-        this._view.setSelections(newSelection);
-        this._view.setFocus(toFocused);
-    }
-
-    /**
-     * @description Applies multi-selection when selecting in single.
-     */
-    private _mutliSelectionInSingle(e: IListMouseEvent<T>): void {
-        const toFocused = e.actualIndex!;
-
-        const currSelection = this._view.getSelections();
-        const newSelection = Arrays.remove(currSelection, toFocused);
-
-        this._view.setFocus(toFocused);
-        this._view.setAnchor(toFocused);
-
-        if (newSelection.length === currSelection.length) {
-            // we are not removing any of the current selections
-            this._view.setSelections([...newSelection, toFocused]);
-        } else {
-            // we removed one of the selections
-            this._view.setSelections(newSelection);
-        }
-    }
-
-    private __getNearestContiguousRange(range: number[], anchor: number): number[] {
-        const index = range.indexOf(anchor);
-        if (index === -1) {
-            return [];
-        }
-
-        const result: number[] = [];
-        let i = index - 1;
-        while (i >= 0 && range[i] === anchor - (index - i)) {
-            result.push(range[i--]!);
-        }
-
-        result.reverse();
-        i = index;
-        while (i < range.length && range[i] === anchor + (i - index)) {
-            result.push(range[i++]!);
-        }
-
-        return result;
-    }
-}
-
-/**
- * @internal
- * @class An internal class that handles the keyboard support of {@link IListWidget}.
- * It handles:
- *  - enter
- *  - up
- *  - down
- *  - page up
- *  - page down
- *  - escape
- */
-class ListWidgetKeyboardController<T> implements IDisposable {
-
-    // [field]
-
-    private _disposables = new DisposableManager();
-    private _view: IListWidget<T>;
-
-    // [constructor]
-
-    constructor(
-        view: IListWidget<T>
-    ) {
-        this._view = view;
-        this._disposables.register(this.onKeydown(e => this.__onDidKeydown(e)));
-    }
-
-    // [private getter]
-
-    @memoize private get onKeydown(): Register<IStandardKeyboardEvent> { return Event.filter(this._view.onKeydown, e => !DomUtility.isInputElement(e.target as HTMLElement)); }
-
-    // [public method]
-
-    public dispose(): void {
-        this._disposables.dispose();
-    }
-
-    // [private helper methods]
-
-    private __onDidKeydown(e: IStandardKeyboardEvent): void {
-        switch (e.key) {
-            case KeyCode.Enter:
-                this.__onEnter(e);
-                break;
-            case KeyCode.UpArrow:
-                this.__onUpArrow(e);
-                break;
-            case KeyCode.DownArrow:
-                this.__onDownArrow(e);
-                break;
-            case KeyCode.PageUp:
-                this.__onPageupArrow(e);
-                break;
-            case KeyCode.PageDown:
-                this.__onPagedownArrow(e);
-                break;
-            case KeyCode.Escape:
-                this.__onEscape(e);
-                break;
-            default:
-                break;
-        }
-    }
-
-    private __onEnter(e: IStandardKeyboardEvent): void {
-        e.preventDefault();
-		e.stopPropagation();
-        const focused = this._view.getFocus();
-        this._view.setSelections(focused !== null ? [focused] : []);
-    }
-
-    private __onUpArrow(e: IStandardKeyboardEvent): void {
-        if (this._view.getFocus() !== null) {
-            e.preventDefault();
-            e.stopPropagation();
-            const newFoused = this._view.focusPrev(1, false, undefined);
-            if (newFoused !== -1) {
-                this._view.setAnchor(newFoused);
-                this._view.reveal(newFoused, undefined);
-            }
-            this._view.setDomFocus();
-        }
-    }
-
-    private __onDownArrow(e: IStandardKeyboardEvent): void {
-        if (this._view.getFocus() !== null) {
-            e.preventDefault();
-            e.stopPropagation();
-            const newFoused = this._view.focusNext(1, false, undefined);
-            if (newFoused !== -1) {
-                this._view.setAnchor(newFoused);
-                this._view.reveal(newFoused, undefined);
-            }
-            this._view.setDomFocus();
-        }
-    }
-
-    private __onPageupArrow(e: IStandardKeyboardEvent): void {
-        e.preventDefault();
-		e.stopPropagation();
-        // TODO
-        console.warn('does not support pageup in ListWidget yet.');
-    }
-
-    private __onPagedownArrow(e: IStandardKeyboardEvent): void {
-        e.preventDefault();
-		e.stopPropagation();
-        // TODO
-        console.warn('does not support pagedown in ListWidget yet.');
-    }
-
-    private __onEscape(e: IStandardKeyboardEvent): void {
-        if (this._view.getSelections().length) {
-            e.preventDefault();
-		    e.stopPropagation();
-            this._view.setSelections([]);
-            this._view.setAnchor(null);
-			this._view.setDomFocus();
-        }
-    }
 }
 
 /**
@@ -742,7 +414,7 @@ export class ListWidget<T> implements IListWidget<T> {
         
         // keyboard support integration
         if (opts.keyboardSupport || opts.mouseSupport === undefined) {
-            const keyboardController = new ListWidgetKeyboardController(this);
+            const keyboardController = this.__createListWidgetKeyboardController();
             this.disposables.register(keyboardController);
         }
         
@@ -929,10 +601,20 @@ export class ListWidget<T> implements IListWidget<T> {
 
     /**
      * @description Creates an instance of a {@link IListWidgetMouseController}.
-     * May override by the inheritance to customize the mouse behaviour.
+     * May override the behaviours by the inheritance to customize the mouse 
+     * behaviour.
      */
     protected __createListWidgetMouseController(opts: IListWidgetOpts<T>): ListWidgetMouseController<T> {
         return new ListWidgetMouseController(this, opts);
+    }
+
+    /**
+     * @description Creates an instance of a {@link ListWidgetKeyboardController}.
+     * May override the behaviours by the inheritance to customize the keyboard
+     * behaviour.
+     */
+    protected __createListWidgetKeyboardController(): ListWidgetKeyboardController<T> {
+        return new ListWidgetKeyboardController(this);
     }
 
     /**
